+++ docs/Changes.md
@@ -6,6 +6,7 @@ Version 0.8.1
 * Proper memory destruction for cuda variables. 
 Some CUDA allocated variables was not properly released at the end of execution.  
 * Minimum gcc version required raised to 7.
+* The scattering pattern memory is allocated during constructor. The `allocate` function is removed.
 
 
 Version 0.8.0